LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/fire-selection-change-on-deleting-empty-element-expected.txt

A


FAIL Selection is updated after removing the element in contenteditable div assert_equals: Character should be removed expected "<h1><br></h1>" but got "<h1>A</h1>"
